Rename action build-scan-setup-maven to setup-maven
