--- build-scan-setup-maven/action.yml
@@ -1,30 +1,0 @@
-name: Maven Build Scan setup action
-description: Setup action to capture Maven Build Scans metadata
-author: Gradle team
-
-inputs:
-  workflow-name:
-    description: 'Name of the workflow triggering the build'
-    required: false
-    default: ${{ github.workflow }}
-  job-name:
-    description: 'Name of the job triggering the build'
-    required: false
-    default: ${{ github.job }}
-  capture-strategy:
-    description: 'Build Scan capture strategy (ALWAYS, ON_FAILURE, ON_DEMAND)'
-    required: false
-    default: ALWAYS
-  capture-unpublished-build-scans:
-    description: 'Whether to enable or not unpublished Build Scans capture'
-    required: false
-    default: true
-  capture-build-scan-links:
-    description: 'Whether to enable or not Build Scan links capture'
-    required: false
-    default: true
-
-runs:
-  using: node20
-  main: ../dist/build-scan-setup-maven/main/index.js
-  post: ../dist/build-scan-setup-maven/post/index.js
