+++ package.json
@@ -22,12 +22,12 @@
     "format:write": "prettier --write */src/**/*.ts",
     "format:test": "prettier --write **/__tests__/**/*.ts",
     "lint": "eslint */src/**/*.ts",
-    "package-build-scan-maven-capture-extension": "mvn clean package -f maven-build-scan-capture-extension/pom.xml",
-    "package-build-scan-maven-setup-main": "ncc build build-scan-setup-maven/src/main.ts --out dist/build-scan-setup-maven/main",
-    "package-build-scan-maven-setup-post": "ncc build build-scan-setup-maven/src/post.ts --out dist/build-scan-setup-maven/post",
-    "package-build-scan-maven-publish": "ncc build maven-publish-build-scan/src/main.ts --out dist/maven-publish-build-scan",
+    "package-maven-capture-extension": "mvn clean package -f maven-build-scan-capture-extension/pom.xml",
+    "package-maven-setup-main": "ncc build maven-setup/src/main.ts --out dist/maven-setup/main",
+    "package-maven-setup-post": "ncc build maven-setup/src/post.ts --out dist/maven-setup/post",
+    "package-maven-publish": "ncc build maven-publish-build-scan/src/main.ts --out dist/maven-publish-build-scan",
     "check": "npm-run-all --parallel format:write format:test lint",
-    "package": "npm-run-all --parallel package-build-scan-*",
+    "package": "npm-run-all --parallel package-maven-*",
     "build": "npm run check && npm run package",
     "test": "jest",
     "all": "npm run build && npm run test"
